+++ migrations/000002_create_tenants_projects.down.sql
@@ -1,0 +1,6 @@
+BEGIN;
+
+DROP TABLE app.projects;
+DROP TABLE app.tenants;
+
+COMMIT;